BEGIN BROKEN PROJECT STATE! Eliminate temporary `DefaultXFormsXPathEvaluator`
As planned, and as described in prior commits where it was introduced, the behavior of this subclass is now restricted to `TestXFormsXPathEvaluator`.
